Create a new crate `encodings` that will have all that is not related to UI. The `encodings_ui` crate will only have UI related components in the future.
