+++ crates/encodings_ui/LICENSE-GPL
