+++ crates/zed/Cargo.toml
@@ -53,7 +53,7 @@ diagnostics.workspace = true
 editor.workspace = true
 zeta2_tools.workspace = true
 encoding.workspace = true
-encodings.workspace = true
+encodings_ui.workspace = true
 env_logger.workspace = true
 extension.workspace = true
 extension_host.workspace = true
