+++ crates/zed/src/main.rs
@@ -630,7 +630,7 @@ pub fn main() {
         zeta::init(cx);
         inspector_ui::init(app_state.clone(), cx);
         json_schema_store::init(cx);
-        encodings::init(cx);
+        encodings_ui::init(cx);
 
         cx.observe_global::<SettingsStore>({
             let http = app_state.client.http_client();
